+++ create_HDR_script.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """Usage:
-    
+
     ./create_HDR_script FIRST-FILENAME
 
 This script takes one argument, the name of the first file to be used in an HDR
@@ -29,41 +29,56 @@ def print_usage(exit_value=0):
     print(__doc__)
     sys.exit(exit_value)
 
-def create_script_from_file_list(HDR_input_files, file_to_move=None, file_to_delete=None):
-    """This procedure creates an enfuse HDR script from a list of files that compose it."""
+def create_script_from_file_list(HDR_input_files, file_to_move=None, file_to_delete=None, delete_originals=False):
+    """This procedure creates an enfuse HDR script from a list of files that compose it.
+    Note that it does not run the script.
+
+    If file_to_move is not None, that file is moved into the old_scripts directory
+    after a successful script creation. If file_to_delete is not None, that file is
+    deleted after a successful script creation.
+
+    The parameter delete_originals controls what the script, when run, does, if
+    successful. If delete_originals is True, the files that have been blended into
+    the tonemapped file are deleted; otherwise, they're moved into a subfolder
+    called HDR_components/. By default, they're moved, not deleted.
+    """
     output_file = os.path.splitext(HDR_input_files[0])[0].strip() + "_HDR.TIFF"
     output_TIFF_base = os.path.splitext(output_file)[0].strip().replace('-','').replace('_','')
-    
+
     the_script = """#!/usr/bin/env bash
 
 # %s from %s ... %s with aligning first
 # This script written by Patrick Mooney's create_HDR_script.py script, see
 #     https://github.com/patrick-brian-mooney/personal-library/blob/master/create_HDR_script.py
 """ % (output_file, HDR_input_files[0], HDR_input_files[-1])
-    
+
     the_script = the_script + """
 align_image_stack -xyzdivvv -a HDR_AIS_%s %s
 enfuse "$@"  --output=%s HDR_AIS_%s*
 rm HDR_AIS_%s*
 """ % (output_TIFF_base, ' '.join(HDR_input_files), output_file, output_TIFF_base, output_TIFF_base)
-    
+
     the_script = the_script + """
 convert %s -quality 95 %s.JPG
 rm %s
 exiftool -tagsfromfile %s %s.JPG
 """ % (output_file, os.path.splitext(output_file)[0], output_file, HDR_input_files[0], os.path.splitext(output_file)[0])
-    
+
     the_script = the_script + """
 rm %s.JPG_original
-mv %s HDR_components/
-""" % (os.path.splitext(output_file)[0], ' '.join(HDR_input_files))
-    
+""" % os.path.splitext(output_file)[0]
+
+    if delete_originals:
+        the_script = the_script + "rm %s" % ' '.join(HDR_input_files)
+    else:
+        the_script = the_script + "mv %s HDR_components/" % ' '.join(HDR_input_files)
+
     script_file_name = os.path.splitext(output_file)[0] + '.SH'
     with open(script_file_name, mode='w') as script_file:
         script_file.write(''.join(the_script))
-    
+
     os.chmod(script_file_name, os.stat(script_file_name).st_mode | 0o111)    # or, as they say in Bash, "chmod a+x SCRIPT_FILE_NAME"
-    
+
     if file_to_move:
         try:
             if not os.path.exists('old_scripts/'):
@@ -83,18 +98,18 @@ def create_script_from_first_file(first_file, num_files=total_number_of_files, f
     """This script creates an enfuse HDR script from the first file on the list and,
     optionally, the number of files that sequentially follow the first file that
     should be input files for the enfuse operation.
-    
+
     To adjust the default number of files to use as inputs to the enfuse operation,
     change the value of the total_number_of_files constant, above.
-    
+
     FILE_TO_DELETE specifies the name of a file (e.g., an old script, for instance)
     to delete if the creation of the new script is successful.
     """
     oldpath = os.getcwd()
     newdir, first_file = os.path.split(first_file)
-    if newdir: 
+    if newdir:
         os.chdir(newdir)
-    
+
     if debugging:
         print('creating script starting with file %s.' % first_file)
         print('     current directory is %s.' % os.getcwd())
@@ -102,12 +117,12 @@ def create_script_from_first_file(first_file, num_files=total_number_of_files, f
     files_in_directory = sorted(glob.glob('*jpg') + glob.glob('*JPG'))
     selected_file_position = files_in_directory.index(os.path.split(first_file)[1])
     HDR_input_files = files_in_directory[selected_file_position : selected_file_position + num_files]
-    
+
     if debugging:
         print('     files in use are: %s' % ' '.join(HDR_input_files))
 
     create_script_from_file_list(HDR_input_files, file_to_delete=file_to_delete)
-            
+
     os.chdir(oldpath)
 
 if __name__ == "__main__":
@@ -117,6 +132,6 @@ if __name__ == "__main__":
     else:
         print("ERROR: You must specify the first file in the sequence.\n")
         print_usage(exit_value=1)
-    
+
     create_script_from_first_file(sys.argv[1])